+++ src/gossamer/core/inducer.clj
@@ -19,15 +19,42 @@
     [gossamer.core.keydef :as kdef]))
 
 
-(defn log-mdc-codec-init
+(defn log-mdc-codec-init-only
   "Initialize SLF4j MDC codec for Logback."
   [context]
   ;; lazily load the namespace to avoid Logback auto-initialization during namespace load
-  (require 'cambium.logback.json.flat-layout)
-  (if-let [f (find-var 'cambium.logback.json.flat-layout/set-decoder!)]
-    (f codec/destringify-val)
-    (throw (ex-info "Cannot find fn 'cambium.logback.json.flat-layout/set-decoder!' in classpath." {})))
+  (core-util/let-var [set-decoder!  'cambium.logback.json.flat-layout/set-decoder!
+                      find-context  'cambium.logback.core.util/find-logger-context
+                      context-name  'cambium.logback.core.util/logger-context-name
+                      start-context 'cambium.logback.core.util/start-logger-context]
+    (set-decoder! codec/destringify-val)
+    (let [logback-context (find-context)]
+      (echo/echof "Starting Logback context '%s'" (context-name logback-context))
+      (start-context logback-context)))
   context)
+
+
+(defn register-logback-deinit
+  "Register a deinit task to flush/clean up the Logback context."
+  [context]
+  (as-> (core-kdef/ctx-deinit context) $
+    (vec $)
+    (conj $ (fn []
+              (core-util/let-var [find-context 'cambium.logback.core.util/find-logger-context
+                                  context-name 'cambium.logback.core.util/logger-context-name
+                                  stop-context 'cambium.logback.core.util/stop-logger-context]
+                (let [logback-context (find-context)]
+                  (echo/echof "Stopping Logback context '%s'" (context-name logback-context))
+                  (stop-context logback-context)))))
+    (assoc context (key core-kdef/ctx-deinit) $)))
+
+
+(defn log-mdc-codec-init
+  "Initialize Logback and register a corresponding deinit callback."
+  [context]
+  (-> context
+    log-mdc-codec-init-only
+    register-logback-deinit))
 
 
 (defn calfpath-routes->ring-handler